% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_criteria_score_plots.R
\name{costs_piechart}
\alias{costs_piechart}
\title{Plot Pie Chart of Survey Total Costs}
\usage{
costs_piechart(smart_tool_output, save_output = FALSE, save_filepath = NULL)
}
\arguments{
\item{smart_tool_output}{A data.frame containing SMART tool inputs and outputs including survey priority scores and ranks returned by the \code{imp_smart_tool} function.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}
}
\value{
Pie chart plot
}
\description{
Plot Pie Chart of Survey Total Costs
}
