% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imp_smart_tool}
\alias{imp_smart_tool}
\title{Calculate Survey Priority Scores with SMART Tool}
\usage{
imp_smart_tool(
  refuge_code = NULL,
  start_year = NULL,
  nyears = NULL,
  annual_staff_budget = NULL,
  survey_scores_filepath = NULL,
  criteria_weights_filepath = NULL,
  save_output = FALSE,
  save_filepath = NULL
)
}
\arguments{
\item{refuge_code}{Refuge 3-character organization code.}

\item{start_year}{Start year of the IMP.}

\item{nyears}{Duration of IMP.}

\item{annual_staff_budget}{Annual total number of FTE weeks available for surveys.}

\item{survey_scores_filepath}{Full file path (including file name) to CSV file containing raw survey scores for SMART tool.}

\item{criteria_weights_filepath}{Full file path (including file name) to CSV file containing criteria weights for SMART tool.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}
}
\value{
A data.frame containing SMART tool inputs and outputs including survey priority scores and ranks.
}
\description{
Calculate Survey Priority Scores with SMART Tool
}
