% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_criteria_score_plots.R
\name{plot_scores}
\alias{plot_scores}
\title{}
\usage{
plot_scores(matplot, mar = c(8, 14, 4, 4))
}
\arguments{
\item{matplot}{Value returned from \code{plot.matrix} package.}

\item{mar}{Argument for margins passed on to \code{plot} function.}
}
\value{
Plot of matrix
}
\description{

}
