% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_output_plots.R
\name{portfolio_grid_plot}
\alias{portfolio_grid_plot}
\title{Plot Top Performing Survey Portfolios Given Scenario Constraints}
\usage{
portfolio_grid_plot(
  scenario_output,
  ndisplay = 10,
  save_output = FALSE,
  save_filepath = NULL
)
}
\arguments{
\item{scenario_output}{A data.frame of ranked portfolios for a given scenario returned by the \code{scenario_optimization_tool} function.}

\item{ndisplay}{Number of portfolios to display. Default is top 10.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}
}
\value{
Grid plot of top-peforming survey portfolios.
}
\description{
Plot Top Performing Survey Portfolios Given Scenario Constraints
}
