% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scenario_optimization_tool}
\alias{scenario_optimization_tool}
\title{Impose Decision Constraints and Optimize Portfolio Selection}
\usage{
scenario_optimization_tool(
  optim_scenario,
  smart_tool_output,
  save_output = FALSE,
  save_filepath = NULL,
  max_output = 10000,
  create_scatterplots = FALSE,
  save_scatterplots = FALSE
)
}
\arguments{
\item{optim_scenario}{List of length 2 with one element containing survey selection constraints and the other containing a logical argument enforcing a budget constraint.}

\item{smart_tool_output}{A data.frame containing SMART tool inputs and outputs including survey priority scores and ranks produced by the \code{imp_smart_tool} function.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}

\item{max_output}{Maximum number of portfolios to output. Default = 10000.}

\item{create_scatterplots}{Logical argument controlling whether to print scatterplots.}

\item{save_scatterplots}{Logical argument controlling whether to save scatterplots. \code{save_filepath} argument used to set save location.}
}
\value{
A data.frame containing survey portfolios meeting scenario constraints and listed in ranking order based on cost-benefit scores.
}
\description{
Impose Decision Constraints and Optimize Portfolio Selection
}
