% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_criteria_score_plots.R
\name{survey_scores_hists}
\alias{survey_scores_hists}
\title{Plot Histograms of Normalized and Weighted Survey Scores}
\usage{
survey_scores_hists(
  smart_tool_output,
  save_output = FALSE,
  save_filepath = NULL
)
}
\arguments{
\item{smart_tool_output}{A data.frame containing SMART tool inputs and outputs including survey priority scores and ranks returned by the \code{imp_smart_tool} function.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}
}
\value{
Histogram plot
}
\description{
Plot Histograms of Normalized and Weighted Survey Scores
}
