% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_criteria_score_plots.R
\name{weighted_score_gridplot}
\alias{weighted_score_gridplot}
\title{Plot Weighted Normalized Survey Scores as Color-Coded Matrix}
\usage{
weighted_score_gridplot(
  smart_tool_output,
  save_output = FALSE,
  save_filepath = NULL
)
}
\arguments{
\item{smart_tool_output}{A data.frame containing SMART tool inputs and outputs including survey priority scores and ranks returned by the \code{imp_smart_tool} function.}

\item{save_output}{Logical value controlling whether output is saved as CSV. Default is FALSE.}

\item{save_filepath}{File path to folder where output is saved. File names are internally generated.}
}
\value{
Plot of matrix
}
\description{
Plot Weighted Normalized Survey Scores as Color-Coded Matrix
}
